# docs/swagger.yaml
openapi: 3.0.3
info:
  title: CrowdFlow Event Ticketing API
  description: |
    API documentation for the CrowdFlow Venue and Event Management platform.

    ### Security Architecture:
    - **Session Authentication**: A short-lived JWT access token is set as a secure `HttpOnly` cookie named `access_token` (Path `/`). It is paired with a long-lived, rotating, server-revocable `refresh_token` cookie (scoped to Path `/api/v1/auth`) which `POST /api/v1/auth/refresh` exchanges for a fresh access token once the access token expires.
    - **CSRF Protection**: State-changing endpoints require Double-Submit Cookie verification. The client extracts the value of the `csrf_token` cookie and submits it in the custom header `X-CSRF-Token`. Only `/api/v1/auth/login` and `/api/v1/auth/register` are exempt.
    - **Admin Console**: Every `/api/v1/admin/*` route requires the `Super Admin` platform role.

    All responses (success and error) are wrapped in a standard JSON response envelope.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: /
    description: Proxy Gateway (Nginx)
paths:
...
  /api/v1/auth/register:
    post:
      summary: User Registration
      description: Registers a new user native credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: User registered successfully
        '400':
          description: Invalid request body (`BAD_REQUEST`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed - missing required fields (`VALIDATION_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email already exists (`EMAIL_ALREADY_REGISTERED`) or other registration failure (`REGISTRATION_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      summary: User Login
      description: |
        Native credentials login. Generates access session cookies.

        Rate limited per client IP: at most 10 attempts per 15-minute window
        (fixed-window counter in Redis; fails open on Redis outage).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication. Sets `access_token`, `refresh_token`, and `csrf_token` cookies. No token is returned in the JSON body.
          headers:
            Set-Cookie:
              description: |
                Sets three cookies:
                1. `access_token`: Short-lived HttpOnly JWT access token (Path `/`).
                2. `refresh_token`: Long-lived HttpOnly rotating refresh token (Path `/api/v1/auth`).
                3. `csrf_token`: Client-readable cookie for CSRF double-submission (Path `/`).
              schema:
                type: string
                example: access_token=eyJhbGciOi...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user_id:
                            type: string
                            example: "12"
                          email:
                            type: string
                            format: email
                            example: admin@crowdflow.com
                          full_name:
                            type: string
                            example: Richie Obhasa
                          role:
                            type: string
                            description: Highest-privilege platform-wide role held by the user.
                            enum: [User, Event Organizer, Super Admin]
                            example: Super Admin
        '400':
          description: Invalid JSON body (`BAD_REQUEST`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid credentials (`AUTHENTICATION_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded - more than 10 login attempts from this IP within 15 minutes (`RATE_LIMITED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      summary: Refresh Session
      description: |
        Rotates the refresh token and issues a new access token. Authenticated
        purely via the `refresh_token` cookie (the `access_token` may be expired,
        so `CookieAuth` is deliberately NOT required here). CSRF-protected like any
        state-changing POST.

        On success the refresh token is rotated and all three auth cookies
        (`access_token`, `refresh_token`, `csrf_token`) are re-set. On any failure â€”
        missing, expired, revoked, or replayed (reuse-detected) token â€” all auth
        cookies are cleared and 401 is returned so the client falls back to a full
        login.

        Rate limited per client IP: at most 60 attempts per 15-minute window
        (fixed-window counter in Redis; fails open on Redis outage).
      security:
        - RefreshCookie: []
        - CSRFToken: []
      responses:
        '200':
          description: Session refreshed. Rotates and re-sets the `access_token`, `refresh_token`, and `csrf_token` cookies.
          headers:
            Set-Cookie:
              description: Re-sets all three auth cookies with a freshly minted access token and rotated refresh token.
              schema:
                type: string
                example: access_token=eyJhbGciOi...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user_id:
                            type: string
                            example: "12"
                          email:
                            type: string
                            format: email
                            example: admin@crowdflow.com
                          full_name:
                            type: string
                            example: Richie Obhasa
                          role:
                            type: string
                            description: Highest-privilege platform-wide role held by the user.
                            enum: [User, Event Organizer, Super Admin]
                            example: Super Admin
        '401':
          description: Refresh token missing, expired, revoked, or replayed (`UNAUTHORIZED`). All auth cookies are cleared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF verification missing or failed (`CSRF_TOKEN_MISSING`, `CSRF_HEADER_MISSING`, `CSRF_TOKEN_MISMATCH`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded - more than 60 refresh attempts from this IP within 15 minutes (`RATE_LIMITED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/google/callback:
    get:
      summary: Google OAuth Callback
      description: Processes Google authorization code, verifies state, provisions user, and redirects the browser based on the user's platform role.
      parameters:
        - name: state
          in: query
          required: true
          description: OAuth state verification token matching the `oauth_state` cookie.
          schema:
            type: string
        - name: code
          in: query
          required: true
          description: Authorization code returned from Google.
          schema:
            type: string
      security:
        - StateCookie: []
      responses:
        '307':
          description: |
            Redirect outcome. On successful login the `oauth_state` cookie is cleared, the `access_token`,
            `refresh_token`, and `csrf_token` session cookies are set, and the browser is redirected by platform role:
            - `Super Admin` â†’ `/admin`
            - `Auditor` â†’ `/auditor`
            - `Event Organizer` â†’ `/dashboard`
            - everyone else â†’ `/`

            If the email is already registered with native credentials, no session is created and the browser
            is redirected to `/login?error=PROVIDER_MISMATCH&provider=native`.
          headers:
            Location:
              schema:
                type: string
                example: /dashboard
        '400':
          description: State verification cookie missing or auth code missing (`BAD_REQUEST`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: State token verification mismatch (`FORBIDDEN`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication code exchange failed (`AUTHENTICATION_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      summary: User Logout
      description: |
        Logs out the current session. Revokes the server-side refresh session (using the
        `refresh_token` cookie) so the refresh token can no longer be used, then clears all
        three auth cookies (`access_token`, `refresh_token`, `csrf_token`). Does not require an
        authenticated access token, but does require CSRF verification (double-submit cookie).
        Logout still succeeds even if server-side revocation fails.
      security:
        - CSRFToken: []
      responses:
        '200':
          description: Successfully logged out. Revokes the server-side refresh session and clears the `access_token`, `refresh_token`, and `csrf_token` cookies.
          headers:
            Set-Cookie:
              description: Overwrites all three auth cookies with MaxAge=-1 to delete them.
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Logged out successfully
        '403':
          description: CSRF verification missing or failed (`CSRF_TOKEN_MISSING`, `CSRF_HEADER_MISSING`, `CSRF_TOKEN_MISMATCH`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout-all:
    post:
      summary: Logout All Devices
      description: |
        Revokes every active session for the authenticated user (all devices), e.g. after a
        suspected account compromise, then clears the current browser's auth cookies. Requires
        an authenticated access token and CSRF verification.
      security:
        - CookieAuth: []
        - CSRFToken: []
      responses:
        '200':
          description: All sessions revoked. Clears the current browser's `access_token`, `refresh_token`, and `csrf_token` cookies.
          headers:
            Set-Cookie:
              description: Overwrites all three auth cookies with MaxAge=-1 to delete them.
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Logged out of all devices
        '401':
          description: Access token missing, invalid, or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF verification missing or failed (`CSRF_TOKEN_MISSING`, `CSRF_HEADER_MISSING`, `CSRF_TOKEN_MISMATCH`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to revoke sessions (`INTERNAL_SERVER_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      summary: Get Current User Profile
      description: Returns the authenticated user's full profile, including stats and associated events (attended events for buyers, organized events for organizers/admins).
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Profile retrieved successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Session token missing, invalid, or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user no longer exists (`NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load profile stats or associated events (`INTERNAL_SERVER_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update Current User Profile
      description: Updates the authenticated user's editable profile fields.
      security:
        - CookieAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Profile updated successfully
        '400':
          description: Invalid JSON body or payload over 1MB (`BAD_REQUEST`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session token missing, invalid, or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF verification missing or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'Validation failed (`VALIDATION_FAILED`): full_name required and â‰¤255 chars, phone_number â‰¤50, location â‰¤255, bio â‰¤2000.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Database update failed (`INTERNAL_SERVER_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Short-lived JWT access token. Sent on every request (Path `/`).
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: |
        Long-lived, rotating, server-revocable refresh token. Scoped to Path
        `/api/v1/auth`, so it is only sent to `/api/v1/auth/refresh` (to rotate)
        and `/api/v1/auth/logout` (to revoke). Used by `POST /api/v1/auth/refresh`
        to mint a new access token when the access token has expired.
    StateCookie:
      type: apiKey
      in: cookie
      name: oauth_state
      description: Google oauth flow verification token.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-Submit CSRF Token header. Must match values in the `csrf_token` cookie.
  schemas:
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Set to true if request completed successfully.
          example: true
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          example: richie@crowdflow.com
        password:
          type: string
          format: password
          example: password123
        full_name:
          type: string
          example: Richie Obhasa
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@crowdflow.com
        password:
          type: string
          format: password
          example: admin123
    UpdateProfileRequest:
      type: object
      required:
        - full_name
      properties:
        full_name:
          type: string
          maxLength: 255
          example: Richie Obhasa
        phone_number:
          type: string
          maxLength: 50
          example: "+62 812 3456 7890"
        location:
          type: string
          maxLength: 255
          example: Jakarta, Indonesia
        bio:
          type: string
          maxLength: 2000
          example: Live music enthusiast.
...